♻️ 모달 중복되는 부분 컴포넌트 분리
